--- codebase/resnet101/config.json
@@ -1,26 +1,0 @@
-{ "version": "baseline152_pt",
-  "batch_size": 125,
-  "hp_random_search": false,
-  "notes": "ResNet152 baseline testing",
-  "pre-format": {
-    "targets": [
-      "id",
-      "config",
-      "optimizer",
-      "test_result_filename"
-    ]
-  },
-  "models": [
-    {
-        "id": "ResNet152 baseline_pt",
-        "model": "rn.resnet152",
-        "config": "v38-4 model152(all)_unfrozen-full (lr=7e-7,wd=4e-8)_{}.config",
-        "last_layer": "Linear({}, {}, bias=False)",
-        "layers_to_unfreeze": ["conv1","bn1","layer1","layer2","layer3","layer4","fc"],
-        "loss": "L1Loss",
-        "optimizer": "torch.optim.Adam(model.parameters(), lr=7e-7, weight_decay=4e-8)",
-        "targets": [0,1,2,3,4,5,6],
-        "test_result_filename": "test_results_baseline152_pt.json"
-    }
-  ]
-}
