--- codebase/resnet101/metrics_v3.json
@@ -1,25 +1,0 @@
-{ 
-    "metrics": [
-        {
-            "metric_id": "b_score"
-        }, 
-        {
-            "metric_id": "rouge2"
-        },
-        {
-            "metric_id": "rouge3"
-        },
-        {
-            "metric_id": "rougeL"
-        },
-        {
-            "metric_id": "meteor"
-        }, 
-        {
-            "metric_id": "bert"
-        },
-        {
-            "metric_id": "cider"
-        }
-    ]
-}
