--- codebase/resnet101/notes.txt
@@ -1,17 +1,0 @@
-AlexNet layers: 
-
-features.0.weight
-features.0.bias
-features.3.weight
-features.3.bias 
-features.6.weight
-features.6.bias
-features.8.weight
-features.8.bias
-features.10.weight
-features.10.bias
-classifier.1.weight
-classifier.1.bias
-classifier.4.weight
-classifier.4.bias
-classifier.6.weight
