--- codebase/resnet101/tmux-client-3136.log
@@ -1,51 +1,0 @@
-1610990163.875681 client started (3136): version 3.1, socket /tmp/tmux-29784/default, protocol 8
-1610990163.875735 on Linux 4.15.0-128-generic #131~16.04.1-Ubuntu SMP Wed Dec 9 17:33:47 UTC 2020; libevent 2.1.10-stable (poll)
-1610990163.875799 socket is /tmp/tmux-29784/default
-1610990163.875820 trying connect
-1610990163.875966 add peer 0x555e3f6823f0: 6 ((nil))
-1610990163.876141 sending message 100 to peer 0x555e3f6823f0 (4 bytes)
-1610990163.876159 sending message 101 to peer 0x555e3f6823f0 (15 bytes)
-1610990163.876169 sending message 102 to peer 0x555e3f6823f0 (11 bytes)
-1610990163.876178 sending message 108 to peer 0x555e3f6823f0 (73 bytes)
-1610990163.876188 sending message 104 to peer 0x555e3f6823f0 (0 bytes)
-1610990163.876198 sending message 107 to peer 0x555e3f6823f0 (4 bytes)
-1610990163.876207 sending message 105 to peer 0x555e3f6823f0 (19 bytes)
-1610990163.876216 sending message 105 to peer 0x555e3f6823f0 (20 bytes)
-1610990163.876224 sending message 105 to peer 0x555e3f6823f0 (16 bytes)
-1610990163.876233 sending message 105 to peer 0x555e3f6823f0 (34 bytes)
-1610990163.876241 sending message 105 to peer 0x555e3f6823f0 (14 bytes)
-1610990163.876249 sending message 105 to peer 0x555e3f6823f0 (29 bytes)
-1610990163.876257 sending message 105 to peer 0x555e3f6823f0 (19 bytes)
-1610990163.876265 sending message 105 to peer 0x555e3f6823f0 (14 bytes)
-1610990163.876275 sending message 105 to peer 0x555e3f6823f0 (92 bytes)
-1610990163.876283 sending message 105 to peer 0x555e3f6823f0 (45 bytes)
-1610990163.876293 sending message 105 to peer 0x555e3f6823f0 (11 bytes)
-1610990163.876301 sending message 105 to peer 0x555e3f6823f0 (24 bytes)
-1610990163.876310 sending message 105 to peer 0x555e3f6823f0 (201 bytes)
-1610990163.876318 sending message 105 to peer 0x555e3f6823f0 (33 bytes)
-1610990163.876326 sending message 105 to peer 0x555e3f6823f0 (47 bytes)
-1610990163.876335 sending message 105 to peer 0x555e3f6823f0 (77 bytes)
-1610990163.876343 sending message 105 to peer 0x555e3f6823f0 (17 bytes)
-1610990163.876351 sending message 105 to peer 0x555e3f6823f0 (26 bytes)
-1610990163.876359 sending message 105 to peer 0x555e3f6823f0 (7 bytes)
-1610990163.876367 sending message 105 to peer 0x555e3f6823f0 (41 bytes)
-1610990163.876380 sending message 105 to peer 0x555e3f6823f0 (8 bytes)
-1610990163.876389 sending message 105 to peer 0x555e3f6823f0 (20 bytes)
-1610990163.876397 sending message 105 to peer 0x555e3f6823f0 (18 bytes)
-1610990163.876405 sending message 105 to peer 0x555e3f6823f0 (53 bytes)
-1610990163.876413 sending message 105 to peer 0x555e3f6823f0 (17 bytes)
-1610990163.876421 sending message 105 to peer 0x555e3f6823f0 (65 bytes)
-1610990163.876429 sending message 105 to peer 0x555e3f6823f0 (53 bytes)
-1610990163.876437 sending message 105 to peer 0x555e3f6823f0 (22 bytes)
-1610990163.876445 sending message 105 to peer 0x555e3f6823f0 (32 bytes)
-1610990163.876454 sending message 105 to peer 0x555e3f6823f0 (45 bytes)
-1610990163.876462 sending message 105 to peer 0x555e3f6823f0 (70 bytes)
-1610990163.876470 sending message 106 to peer 0x555e3f6823f0 (0 bytes)
-1610990163.876488 sending message 200 to peer 0x555e3f6823f0 (4 bytes)
-1610990163.876496 client loop enter
-1610990163.878656 peer 0x555e3f6823f0 message 207
-1610990163.878662 sending message 208 to peer 0x555e3f6823f0 (0 bytes)
-1610990168.398000 peer 0x555e3f6823f0 message 201
-1610990168.398038 sending message 205 to peer 0x555e3f6823f0 (0 bytes)
-1610990168.398457 peer 0x555e3f6823f0 message 204
-1610990168.398484 client loop exit
