+++ codebase/resnet_based_icds_predictor/metrics_v3.json
@@ -1,0 +1,25 @@
+{ 
+    "metrics": [
+        {
+            "metric_id": "b_score"
+        }, 
+        {
+            "metric_id": "rouge2"
+        },
+        {
+            "metric_id": "rouge3"
+        },
+        {
+            "metric_id": "rougeL"
+        },
+        {
+            "metric_id": "meteor"
+        }, 
+        {
+            "metric_id": "bert"
+        },
+        {
+            "metric_id": "cider"
+        }
+    ]
+}
